Rock,Paper, and Scissors Game
This is another python beginner project that I did. It is a simple game of rock, paper, and scissors game. 
The program allows the user to play against the computer and when the user is done playing by input q.
After inputting q, the program will output the round number, what the player and computer chose, and the outcome of the round.